requirements.txt
ipykernel
opencv-python
pandas
matplotlib
numpy
pyyaml
omegaconf
python-dotenv
hydra-core